# docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API gateway",
        "title": "Web api gateway API",
        "contact": {
            "name": "ru.system.ru@gmail.com",
            "email": "ru.system.ru@gmail.com"
        },
        "version": "1.0"
    },
    "host": "91.243.71.100:8080",
    "basePath": "/api/web-api-gateway/v1",
    "paths": {
...
        "/supplier": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение всех поставщиков компании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supplier"
                ],
                "summary": "Get suppliers",
                "operationId": "get-suppliers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Supplier"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создание поставщика",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supplier"
                ],
                "summary": "Create supplier",
                "operationId": "create-supplier",
                "parameters": [
                    {
                        "description": "Необходимо указать данные поставщика.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.InputSupplier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.IdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/supplier/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение поставщика по id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supplier"
                ],
                "summary": "Get supplier by id",
                "operationId": "get-supplier",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Supplier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Supplier"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновление поставщика",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supplier"
                ],
                "summary": "Update supplier",
                "operationId": "update-supplier",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Необходимо указать данные поставщика.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateSupplier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаление поставщика своей компании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supplier"
                ],
                "summary": "Delete supplier",
                "operationId": "delete-supplier",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.IdResponse": {
            "type": "object",
            "properties": {
                "id": {}
            }
        },
        "domain.InputSupplier": {
            "type": "object",
            "required": [
                "actual_address",
                "balance",
                "bank_details",
                "contact_person",
                "contract_number",
                "legal_address",
                "name",
                "phone",
                "warehouse_address"
            ],
            "properties": {
                "actual_address": {
                    "description": "Фактический адрес поставщика",
                    "type": "string",
                    "example": "Фактический адрес поставщика"
                },
                "balance": {
                    "description": "Баланс по поставщику",
                    "type": "number",
                    "example": 100
                },
                "bank_details": {
                    "description": "Банковские реквизиты поставщика",
                    "type": "string",
                    "example": "Банковские реквизиты поставщика"
                },
                "comments": {
                    "description": "Комментарии",
                    "type": "string",
                    "example": "Комментарии"
                },
                "contact_person": {
                    "description": "Контактное лицо у поставщика",
                    "type": "string",
                    "maxLength": 140,
                    "example": "Контактное лицо у поставщика"
                },
                "contract_number": {
                    "description": "Номер и дата договора с поставщиком",
                    "type": "string",
                    "example": "Номер и дата договора с поставщиком"
                },
                "country": {
                    "description": "Страна поставщика",
                    "type": "string",
                    "example": "Страна поставщика"
                },
                "email": {
                    "description": "Электронная почта поставщика",
                    "type": "string",
                    "example": "Электронная почта поставщика"
                },
                "files": {
                    "description": "Ссылки на файлы или документы",
                    "type": "string",
                    "example": "Ссылки на файлы или документы"
                },
                "is_active": {
                    "description": "Статус активности поставщика (активен/неактивен)",
                    "type": "boolean",
                    "example": true
                },
                "legal_address": {
                    "description": "Юридический адрес поставщика",
                    "type": "string",
                    "example": "Юридический адрес поставщика"
                },
                "name": {
                    "description": "Наименование поставщика",
                    "type": "string",
                    "maxLength": 140,
                    "minLength": 5,
                    "example": "ООО Название поставщика"
                },
                "other_fields": {
                    "description": "Дополнительные пользовательские поля",
                    "type": "object",
                    "additionalProperties": true
                },
                "payment_terms": {
                    "description": "Условия оплаты по контракту",
                    "type": "string",
                    "example": "Условия оплаты по контракту"
                },
                "phone": {
                    "description": "Телефон поставщика",
                    "type": "string",
                    "minLength": 5,
                    "example": "Телефон поставщика"
                },
                "product_categories": {
                    "description": "Категории товаров, поставляемых поставщиком",
                    "type": "string",
                    "example": "Категории товаров, поставляемых поставщиком"
                },
                "product_types": {
                    "description": "Количество типов товаров от поставщика",
                    "type": "integer",
                    "example": 100
                },
                "purchase_amount": {
                    "description": "Общая сумма закупок у поставщика",
                    "type": "number",
                    "example": 100
                },
                "region": {
                    "description": "Регион или штат поставщика",
                    "type": "string",
                    "example": "Регион или штат поставщика"
                },
                "registration_date": {
                    "description": "Дата регистрации поставщика",
                    "type": "string",
                    "example": "2022-01-01T00:00:00Z"
                },
                "tax_id": {
                    "description": "Идентификационный номер налогоплательщика (ИНН)",
                    "type": "string",
                    "example": "Идентификационный номер налогоплательщика (ИНН)"
                },
                "warehouse_address": {
                    "description": "Адрес склада поставщика",
                    "type": "string",
                    "example": "Адрес склада поставщика"
                },
                "website": {
                    "description": "Сайт поставщика",
                    "type": "string",
                    "example": "Сайт поставщика"
                }
            }
        },
...
        "domain.Supplier": {
            "type": "object",
            "required": [
                "name",
                "phone"
            ],
            "properties": {
                "actual_address": {
                    "description": "Фактический адрес поставщика",
                    "type": "string"
                },
                "balance": {
                    "description": "Баланс по поставщику",
                    "type": "number"
                },
                "bank_details": {
                    "description": "Банковские реквизиты поставщика",
                    "type": "string"
                },
                "comments": {
                    "description": "Комментарии",
                    "type": "string"
                },
                "company_id": {
                    "description": "Уникальный идентификатор компании",
                    "type": "integer"
                },
                "contact_person": {
                    "description": "Контактное лицо у поставщика",
                    "type": "string"
                },
                "contract_number": {
                    "description": "Номер и дата договора с поставщиком",
                    "type": "string"
                },
                "country": {
                    "description": "Страна поставщика",
                    "type": "string"
                },
                "email": {
                    "description": "Электронная почта поставщика",
                    "type": "string"
                },
                "files": {
                    "description": "Ссылки на файлы или документы",
                    "type": "string"
                },
                "id": {
                    "description": "Уникальный идентификатор поставщика",
                    "type": "integer"
                },
                "is_active": {
                    "description": "Статус активности поставщика (активен/неактивен)",
                    "type": "boolean"
                },
                "legal_address": {
                    "description": "Юридический адрес поставщика",
                    "type": "string"
                },
                "name": {
                    "description": "Наименование поставщика",
                    "type": "string",
                    "maxLength": 140,
                    "minLength": 5
                },
                "other_fields": {
                    "description": "Дополнительные пользовательские поля",
                    "type": "object",
                    "additionalProperties": true
                },
                "payment_terms": {
                    "description": "Условия оплаты по контракту",
                    "type": "string"
                },
                "phone": {
                    "description": "Телефон поставщика",
                    "type": "string"
                },
                "product_categories": {
                    "description": "Категории товаров, поставляемых поставщиком",
                    "type": "string"
                },
                "product_types": {
                    "description": "Количество типов товаров от поставщика",
                    "type": "integer"
                },
                "purchase_amount": {
                    "description": "Общая сумма закупок у поставщика",
                    "type": "number"
                },
                "region": {
                    "description": "Регион или штат поставщика",
                    "type": "string"
                },
                "registration_date": {
                    "description": "Дата регистрации поставщика",
                    "type": "string"
                },
                "tax_id": {
                    "description": "Идентификационный номер налогоплательщика (ИНН)",
                    "type": "string"
                },
                "warehouse_address": {
                    "description": "Адрес склада поставщика",
                    "type": "string"
                },
                "website": {
                    "description": "Сайт поставщика",
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateSupplier": {
            "type": "object",
            "required": [
                "actual_address",
                "balance",
                "bank_details",
                "contact_person",
                "contract_number",
                "legal_address",
                "name",
                "phone",
                "warehouse_address"
            ],
            "properties": {
                "actual_address": {
                    "description": "Фактический адрес поставщика",
                    "type": "string",
                    "example": "Фактический адрес поставщика"
                },
                "balance": {
                    "description": "Баланс по поставщику",
                    "type": "number",
                    "example": 100
                },
                "bank_details": {
                    "description": "Банковские реквизиты поставщика",
                    "type": "string",
                    "example": "Банковские реквизиты поставщика"
                },
                "comments": {
                    "description": "Комментарии",
                    "type": "string",
                    "example": "Комментарии"
                },
                "contact_person": {
                    "description": "Контактное лицо у поставщика",
                    "type": "string",
                    "maxLength": 140,
                    "example": "Контактное лицо у поставщика"
                },
                "contract_number": {
                    "description": "Номер и дата договора с поставщиком",
                    "type": "string",
                    "example": "Номер и дата договора с поставщиком"
                },
                "country": {
                    "description": "Страна поставщика",
                    "type": "string",
                    "example": "Страна поставщика"
                },
                "email": {
                    "description": "Электронная почта поставщика",
                    "type": "string",
                    "example": "Электронная почта поставщика"
                },
                "files": {
                    "description": "Ссылки на файлы или документы",
                    "type": "string",
                    "example": "Ссылки на файлы или документы"
                },
                "is_active": {
                    "description": "Статус активности поставщика (активен/неактивен)",
                    "type": "boolean",
                    "example": true
                },
                "legal_address": {
                    "description": "Юридический адрес поставщика",
                    "type": "string",
                    "example": "Юридический адрес поставщика"
                },
                "name": {
                    "description": "Наименование поставщика",
                    "type": "string",
                    "maxLength": 140,
                    "minLength": 5,
                    "example": "ООО Название поставщика"
                },
                "other_fields": {
                    "description": "Дополнительные пользовательские поля",
                    "type": "object",
                    "additionalProperties": true
                },
                "payment_terms": {
                    "description": "Условия оплаты по контракту",
                    "type": "string",
                    "example": "Условия оплаты по контракту"
                },
                "phone": {
                    "description": "Телефон поставщика",
                    "type": "string",
                    "minLength": 5,
                    "example": "Телефон поставщика"
                },
                "product_categories": {
                    "description": "Категории товаров, поставляемых поставщиком",
                    "type": "string",
                    "example": "Категории товаров, поставляемых поставщиком"
                },
                "product_types": {
                    "description": "Количество типов товаров от поставщика",
                    "type": "integer",
                    "example": 100
                },
                "purchase_amount": {
                    "description": "Общая сумма закупок у поставщика",
                    "type": "number",
                    "example": 100
                },
                "region": {
                    "description": "Регион или штат поставщика",
                    "type": "string",
                    "example": "Регион или штат поставщика"
                },
                "registration_date": {
                    "description": "Дата регистрации поставщика",
                    "type": "string",
                    "example": "2022-01-01T00:00:00Z"
                },
                "tax_id": {
                    "description": "Идентификационный номер налогоплательщика (ИНН)",
                    "type": "string",
                    "example": "Идентификационный номер налогоплательщика (ИНН)"
                },
                "warehouse_address": {
                    "description": "Адрес склада поставщика",
                    "type": "string",
                    "example": "Адрес склада поставщика"
                },
                "website": {
                    "description": "Сайт поставщика",
                    "type": "string",
                    "example": "Сайт поставщика"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}